Chapter Header Anchor Links
When reading a blog post with sub headers it is useful to link to a specific section of the post. As such I have added the ability to enable (off by default) in config anchor link along side the chapter headers in a post. If enabled you will see a link icon which will have an anchor to that chapter heading.

In doing this feature, I have had to alter way the chapter headings are spaced using margins over padding. This allows us to use a hidden margin to push the anchor point to below the nav/menu bar.